refactor ascii ebcdic conversion
