+++ GenevaIO/pom.xml
@@ -58,10 +58,6 @@ under the License.
 			<artifactId>utilities</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>net.sf.jt400</groupId>
-			<artifactId>jt400</artifactId>
-		</dependency>
-		<dependency>
 			<groupId>commons-lang</groupId>
 			<artifactId>commons-lang</artifactId>
 		</dependency>
