+++ PETestFramework/src/main/java/org/genevaers/testframework/TestDriver.java
@@ -27,8 +27,6 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
-import java.nio.CharBuffer;
-import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -824,7 +822,7 @@ public class TestDriver {
               fileIn = new ZFile(dataset,"rb,type=record,recfm=" + recfm.toLowerCase() + ",lrecl=" + lrecl + ",noseek");
               byte[] recBuf = new byte[Integer.parseInt(lrecl)];
               while (fileIn.read(recBuf) != -1) {
-                  ByteBuffer convbb = ByteBuffer.wrap(ebcdicToAscii(recBuf));
+                  ByteBuffer convbb = ByteBuffer.wrap(GersFilesUtils.ebcdicToAscii(recBuf));
                   convbb.position(recBuf.length);
                   fw.writeArray(convbb);
               }
@@ -843,15 +841,4 @@ public class TestDriver {
           }
           return copied;
       }
-
-      private static byte[] ebcdicToAscii(byte[] buf) {
-        Charset utf8charset = Charset.forName("ISO8859-1");
-        Charset ebccharset = Charset.forName(GersConfigration.getZosCodePage());
-        ByteBuffer inputBuffer = ByteBuffer.wrap(buf);
-        CharBuffer data = ebccharset.decode(inputBuffer);
-        return utf8charset.encode(data).array();
-      }
-  
-
-
 }
